Program Participation Observation profile for design elaboration
https://github.com/AuDigitalHealth/ci-fhir-r4/issues/126
Added profile to provide a representation of an individual’s participation in a program, e.g. disease screening or substance use therapy. Participation information may include information on an individual’s eligibility for a program, suspension of participation in a program, or status of participation in a program.
profile and the linked examples demonstrate outcomes of design meetings 02 FEB 2022. These are intended to be used only for design elaboration activities and is reflective of the proposed HL7 AU structure. ADHA profile will derive and constrain to support project-specific NCSR to MHR integration requirements.
Updated cancer screening examples to reflect forecast status (previous examples incorrectly included 'next action')