ImplementationGuide resource - add placeholder resource desriptions
Eliminates warnings 
"Unable to find ImplementationGuide.definition.resource.description for the resource Condition/uti-es-02. Descriptions are strongly encouraged if they cannot be inferred from the resource to allow proper population of the artifact list." and
"Unable to find ImplementationGuide.definition.resource.description for the resource HealthcareService/ncsp. Descriptions are strongly encouraged if they cannot be inferred from the resource to allow proper population of the artifact list."